az-python/requirements.txt
azure-mgmt-resource>=2.0.0
azure-mgmt-compute>=4.0.1
azure-mgmt-network>=1.7.1
azure-mgmt-authorization>=0.50.0
# Only for User Assigned creation
azure-mgmt-msi>=0.2.0
azure-identity>=1.11.0